requirements.txt
# Core dependencies
Django==5.0.1  # Latest stable as of Jan 2024
#redis==5.0.1   # Stable version with good async support
confluent-kafka==2.3.0  # Latest stable with good reliability
pytz==2024.1   # Latest stable
djangorestframework
aioredis>=2.0.0
psycopg
faker
redis[hiredis]>=4.5.0
openai==1.56.2
pytest-django


# Additional dependencies needed
python-dateutil==2.8.2  # For robust date handling
requests==2.31.0  # For notification sender HTTP calls
aiohttp==3.9.1  # For async HTTP requests if needed
typing-extensions==4.11.0  # For better type hints
setuptools==75.6.0

# Development dependencies
black==24.1.1  # Code formatting
flake8==7.0.0  # Linting
mypy==1.8.0   # Type checking
pytest==7.4.4  # Testing
isort==5.13.2
pytest-asyncio==0.23.3  # Async test support
pytest-django==4.7.0  # Django test integration
pytest-mock==3.14.0

# Optional monitoring/metrics
#prometheus-client==0.19.0  # If we want to add metrics later

# For deployment
gunicorn==21.2.0  # WSGI HTTP Server
uvicorn==0.27.0  # ASGI server for async support